// frontend/src/api/generated/models/CreatePromptRatingDto.ts
//@ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * c4 GenAI Suite
 * c4 GenAI Suite
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreatePromptRatingDto
 */
export interface CreatePromptRatingDto {
    /**
     * Rating value from 1 to 5 stars
     * @type {number}
     * @memberof CreatePromptRatingDto
     */
    rating: number;
    /**
     * Optional comment about the prompt
     * @type {string}
     * @memberof CreatePromptRatingDto
     */
    comment?: string;
}

/**
 * Check if a given object implements the CreatePromptRatingDto interface.
 */
export function instanceOfCreatePromptRatingDto(value: object): value is CreatePromptRatingDto {
    if (!('rating' in value) || value['rating'] === undefined) return false;
    return true;
}

export function CreatePromptRatingDtoFromJSON(json: any): CreatePromptRatingDto {
    return CreatePromptRatingDtoFromJSONTyped(json, false);
}

export function CreatePromptRatingDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreatePromptRatingDto {
    if (json == null) {
        return json;
    }
    return {
        
        'rating': json['rating'],
        'comment': json['comment'] == null ? undefined : json['comment'],
    };
}

export function CreatePromptRatingDtoToJSON(value?: CreatePromptRatingDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'rating': value['rating'],
        'comment': value['comment'],
    };
}